Detect mismatching IPC participants with a build-time constant
The IPC protocol is intentionally unversioned, to keep it simple and
cheap, and avoid managing schemas for something that is only used
between strongly-connected components that are deployed together on
one machine. It is *not* the protocol that the clients will talk to
the server.

In development, we might have separately built and run different
components in a more cowboy fashion. I don't want to debug those
differences, so detect them cheaply by exchanging hashes in an initial
handshake.

We used to hash the current executable, and that worked great
when the client and server were both the same executable, like
they are in production.

That failed utterly for integration tests, with custom clients.

With this, we might actually split the `tere-server` executable
into per-service executables, instead of using subcommands.