Stop using async for now
The Rust async ecosystem still gets in the way quite often, and most of
our code runs in narrowly-scoped sandboxes with limited number of
connections. And we still probably want to use threads in some places to
lock the sandbox further based on the task.

The frontmost service, `tere-server`, will still likely use async to
serve HTTP.