Use `tere-sessions` for D-Bus and systemd-machined access control
Previously, we tried to introduce an extra `tere-dbus` group, and use
Systemd `SupplementaryGroups=` to add that where needed. It turns out
that Polkit rules do not see the supplementary groups of the relevant
process, but perform a lookup from `/etc/group` (/NSS/Varlink) instead.

Move access control to work directly with the username `tere-sessions`
instead.

Also fix a copy-pasto in docs where it was claimed that `tere-sessions`
has group `tere-socket-sessions`. That would be pretty pointless,
authorizing it to connect to itself.